requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
build==0.10.0
CacheControl==0.12.14
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
cryptography==41.0.2
distlib==0.3.6
dulwich==0.21.5
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
html5lib==1.1
hydra-core==1.3.2
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
installer==0.7.0
isort==5.10.1
jaraco.classes==3.3.0
jaxtyping==0.2.19
jeepney==0.8.0
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
keyring==23.13.1
lazy-object-proxy==1.7.1
lightning-utilities==0.9.0
lockfile==0.12.2
lpips==0.1.4
Markdown==3.4.3
MarkupSafe==2.1.3
mccabe==0.7.0
more-itertools==9.1.0
msgpack==1.0.5
mypy-extensions==0.4.3
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
pexpect==4.8.0
Pillow==9.5.0
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
protobuf==4.23.3
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyproject_hooks==1.0.0
PyYAML==6.0
rapidfuzz==2.15.1
referencing==0.29.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rpds-py==0.8.10
rsa==4.9
scipy==1.10.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tomli==2.0.1
tomlkit==0.11.8
torch-fidelity==0.3.0
torchinfo==1.8.0
torchmetrics==1.0.3
tqdm==4.65.0
trove-classifiers==2023.7.6
typeguard==2.13.3
typing_extensions==4.6.3
urllib3==1.26.16
virtualenv==20.24.0
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.15.0
